See ChangeLog

+++ Main/Source/iconf.cpp
@@ -1,10 +1,12 @@
 #include "iconf.h"
+#include "felibdef.h"
 #include "game.h"
 #include "feio.h"
 #include "ivandef.h"
 #include "area.h"
 #include "graphics.h"
 #include "bitmap.h"
+
 
 stringoption ivanconfig::DefaultName(	  "DefaultName",
 					  "player's default name",
@@ -169,7 +171,7 @@ ulong ivanconfig::ApplyContrastTo(ulong L)
 
 void ivanconfig::CalculateContrastLuminance()
 {
-  ushort Element = Min((GetContrast() << 7) / 100, 255);
+  ushort Element = Min((GetContrast() << 7) / 100, long(255));
   ContrastLuminance = MakeRGB24(Element, Element, Element);
 }
